refactor: split mesh_generator.py into specific backend modules
